Refactor(changhoon/find_id) : 폴더명 컨벤션 적용 #52
